requirements.txt
torch>=2.0
torch-geometric>=2.4
transformers>=4.30
networkx>=3.0
biopython>=1.80
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
tqdm
python-louvain
requests
